+++ src/Recipe/Step/SaveObject.php
@@ -62,6 +62,7 @@ class SaveObject
                         'parameters' => [
                             'id' => $object->getId(),
                         ],
+                        'formHandleRequest' => false
                     ]);
                 }
             },