+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Teknoo Software - Common - Change Log
 
+## [1.4.6] - 2022-11-25
+### Stable Release
+- `FormHandlingInterface` step is recalled in `EditObjectEndPoint` before re-rendering form when the form has been
+  successfully saved.
+
 ## [1.4.5] - 2022-11-25
 ### Stable Release
 - Update symfony configuration for behat